Add recruitment of custom troops from raided villages and town alleys
T1 troops can be recruited in villages that have previously been raided.
Troops that are not recruited this way, are distributed to roaming bandit parties.
If the player owns an alley in a town they can build a widows' refuge.
The refuge converts available bandit troops in the alley to widow bandit troops.